Infinite scrolling in task list (#8)
Adds infinite scrolling in the task list via a new useInfiniteTasks
query hook. Limit is set to 30 by default.


https://github.com/user-attachments/assets/8f30cb08-8bab-4f4f-96b3-a04e8a8dd528


Currently, the tasks are showing in reverse order due to the default
sort on the server being asc. We should change this to be desc in the
future, but that shouldn't block this PR