Order tasks desc by default (#18)
Changes the TaskRespository's list tasks method to order by desc, by
default. This makes the most recently created tasks appear at the
beginning of the list, which is more intuitive imo and more convenient
for the frontend.

Ideally, we would expose this as a parameter in the `/tasks` endpoint,
but then we should do that change across all bulk get endpoints, and
that would require an SDK change. I think this is okay for now