Rename acp_type=AGENTIC to ASYNC with backwards compatibility (#20)
This PR makes a major refactor to the Agentex server code by adding the
"async" ACP type, which is meant to replace "agentic". New agents should
use "async", but backwards compatibility with "agentic" agents is still
maintained, though marked as deprecated.

In the future, we should try to migrate all existing "agentic" agents to
async, then completely remove it from the codebase.

A couple surface areas that I want to comment on: 
* **Migrations**: A migration to change the default acp type on the
Agents table from "agentic" to "async" has been added
* **Docs**: All references to "agentic" in docs have been removed and
replaced with "async".
* **Testing**: All existing tests using the "agentic" acp type have been
updated to use the new "async" designation, and all of them are passing.
New unit and integration tests have also been added to test backwards
compatibility.
* **Agentex-ui**: No updates have been made to agentex-ui yet, we need
to wait for the new acp type to be available via the SDK after this PR
lands.
* **python SDK**: A separate PR will need to be made (in addition to the
automated stainless ones) to update the jinja templates in the python
SDK to make async the new default.