Add high limit to bulk message endpoint (#83)
Replaces the agentex-sdk's implementation of subscribeTaskState with a
custom, near identical version that includes a 1000 limit on the
messages endpoint.

This is intended to be a stop gap solution to an issue where only the
last 50 messages appear in the UI. The long term fix will be to add an
order_by field to our bulk get endpoints and changing the typescript sdk
to order the messages by desc. This PR will be effectively reverted once
that is completed.